workspace/fiducial/tracking/requirements.txt
qrcode==5.1
pillow==2.7
zbar==0.10
